% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RMATGEset.R
\name{RMATGEset}
\alias{RMATGEset}
\title{RMA normalize TG-Gate affymetrix microarray dataset}
\usage{
RMATGEset(data_dir, anno_db = "rat2302")
}
\arguments{
\item{anno_db}{Annotation database for microarray Defaults to "rat2302"}
}
\description{
Analyze the TG-Gate dataset under the supplied folder. Suppose the .CEL files are in "celfiles" subfolder. The experimental metadata is in "Attribute.tsv" file.
}

